Fixed pages, components and schema compared 2 main
